[5] develop default simulation

+++ WarehouseData/WarehouseNode.cs
@@ -1,3 +1,3 @@
 namespace FF.WarehouseData;
 
-public record WarehouseNode(NodeType Type, (double CenterX, double CenterY) Coordinates);
+public record WarehouseNode(NodeType Type, (int CenterX, int CenterY) Coordinates);
